requirements.txt
Flask==0.11.1
feedgen==0.4.0
beautifulsoup4==4.5.1
pytz==2016.6.1
requests==2.4.3
GoogleAppEngineCloudStorageClient==1.9.22.1
